move remaining text_generation stuff to _common
