+++ src/huggingface_hub/inference/_generated/_async_client.py
@@ -54,6 +54,7 @@ from huggingface_hub.inference._common import (
     _is_tgi_server,
     _open_as_binary,
     _set_as_non_tgi,
+    raise_text_generation_error,
 )
 from huggingface_hub.inference._generated.types import (
     AudioClassificationOutputElement,
@@ -81,9 +82,6 @@ from huggingface_hub.inference._generated.types import (
     ZeroShotImageClassificationOutputElement,
 )
 from huggingface_hub.inference._templating import render_chat_prompt
-from huggingface_hub.inference._text_generation import (
-    raise_text_generation_error,
-)
 from huggingface_hub.inference._types import (
     ConversationalOutput,  # soon to be removed
 )
